pyproject.toml
[project]
name = "myweepal"
version = "2.0.0"
description = "Local conversational life-story AI for Apple Silicon"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "David Irvine", email = "david@maidsafe.com" }
]

dependencies = [
    # Core ML Framework
    "mlx>=0.21.0",
    "mlx-lm>=0.20.0",
    # Speech Recognition
    # Note: mlx-whisper has incompatible dependencies for Python 3.11
    # Will install separately or use alternative
    # Audio Processing
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    "librosa>=0.10.1", # Newer version for Python 3.11 compatibility
    "audiofile>=1.3.0",
    "torchaudio>=2.0.0",
    "numba>=0.57.0", # Force newer version for Python 3.11
    # Vision & Emotion
    "opencv-python>=4.8.0",
    # fer and face-recognition will be installed separately due to dependency conflicts
    # Vector Database & Memory
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.0",
    # API & Web Interface
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "gradio>=4.0.0",
    "websockets>=12.0",
    # Utilities
    "numpy>=1.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.6",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
# Optional dependencies for advanced features
emotion = ["fer>=22.4.0"]
whisper = ["openai-whisper>=20231117"]
tts = ["TTS>=0.22.0"]
speaker = ["speechbrain>=1.0.0", "pyannote-audio>=3.1.0"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
